+++ spec/dummy/db/migrate/20200717132037_add_grant_type_to_access_tokens.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+class AddGrantTypeToAccessTokens < ActiveRecord::Migration[6.0]
+  def change
+    add_column(
+      :oauth_access_tokens,
+      :grant_type,
+      :string,
+      null: true,
+    )
+  end
+end