+++ spec/lib/oauth/client_credentials/creator_spec.rb
@@ -18,6 +18,11 @@ RSpec.describe Doorkeeper::OAuth::ClientCredentials::Creator do
     end.to change { Doorkeeper::AccessToken.count }.by(1)
   end
 
+  it "assigns the client_credentials grant type to the new token" do
+    creator.call(client, scopes)
+    expect(Doorkeeper::AccessToken.last.grant_type).to eq("client_credentials")
+  end
+
   context "when reuse_access_token is true" do
     before do
       allow(Doorkeeper.config).to receive(:reuse_access_token).and_return(true)